HDFS connection is working now, however values are completely hardcoded.
I also had to add ELF libraries in order to make libomptarget believe we
are actually loading the binary. It checks the size of the offload table
to see if offloading really did work. Maybe we need to come up with a way
to "fake" offloading while still having a valid offload table.