Added initial code for mapping addresses to HDFS files.
Basically, each passed address will be mapped to a file. Just before a
program has to be run, the most up-to-date address table is written into
a special file, which has to be read by the kernel.

Currently, there is no way to know whether an address is input or
output given the table, so the kernel doesn't know what to do. We should
define how to store inputs and outputs before diving more into this, I
believe.